microservices/quiz-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
kafka-python==2.0.2
boto3==1.34.0
gunicorn==21.2.0
PyPDF2==3.0.1
